+++ packages/form-schema/src/fields/dynamic/schema.ts
@@ -6,6 +6,41 @@ import type {
 	PhaseType,
 } from "../../core";
 
+export type DynamicGetSchemaParams<
+	FieldSchema extends FieldSchemaBase,
+	Values extends Record<string, any> = Record<string, any>,
+	RawValues extends Record<string, any> = Record<string, any>,
+	SerializedValues extends Record<string, any> = Record<string, any>,
+	Errors extends Record<string, any> = Record<string, any>,
+> = {
+	/**
+	 * object of values of form, type depends from `phase` (2nd argument)
+	 */
+	values: Values | RawValues;
+	/**
+	 * current phase
+	 */
+	phase: PhaseType;
+	/**
+	 * current `getFieldSchema`
+	 */
+	getFieldSchema: GetFieldSchema<FieldSchema>;
+	/**
+	 * global `getFieldType`
+	 */
+	getFieldType: GetFieldType<
+		FieldSchema,
+		Values,
+		RawValues,
+		SerializedValues,
+		Errors
+	>;
+	/**
+	 * stack of parent fields above current field with runtime values
+	 */
+	parents: readonly ParentType[];
+};
+
 export type DynamicSchema<
 	FormApi,
 	FieldSchema extends FieldSchemaBase,
@@ -18,64 +53,26 @@ export type DynamicSchema<
 	 * Callback that should return schema of field or `null` if field can't be shown
 	 */
 	getSchema: (
-		/**
-		 * object of values of form, type depends from `phase` (2nd argument)
-		 */
-		values: Values | RawValues,
-		/**
-		 * current phase
-		 */
-		phase: PhaseType,
-		/**
-		 * current `getFieldSchema`
-		 */
-		getFieldSchema: GetFieldSchema<FieldSchema>,
-		/**
-		 * global `getFieldType`
-		 */
-		getFieldType: GetFieldType<
+		params: DynamicGetSchemaParams<
 			FieldSchema,
 			Values,
 			RawValues,
 			SerializedValues,
 			Errors
 		>,
-		/**
-		 * stack of parent fields above current field with runtime values
-		 */
-		parents: readonly ParentType[],
 	) => FieldSchema | null;
 
 	/**
 	 * Async callback that can be used instead of `getSchema` on parsing phase
 	 */
 	getSchemaAsync?: (
-		/**
-		 * object of values of form, type depends from `phase` (2nd argument)
-		 */
-		values: Values | RawValues,
-		/**
-		 * current phase
-		 */
-		phase: PhaseType,
-		/**
-		 * current `getFieldSchema`
-		 */
-		getFieldSchema: GetFieldSchema<FieldSchema>,
-		/**
-		 * global `getFieldType`
-		 */
-		getFieldType: GetFieldType<
+		params: DynamicGetSchemaParams<
 			FieldSchema,
 			Values,
 			RawValues,
 			SerializedValues,
 			Errors
 		>,
-		/**
-		 * stack of parent fields above current field with runtime values
-		 */
-		parents: readonly ParentType[],
 	) => Promise<FieldSchema>;
 
 	/**